# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ai/models": {
      "get": {
        "operationId": "AiController_getModels",
        "parameters": [],
        "responses": {
          "200": {
            "description": "模型目录列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelCatalogItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取可切换的模型目录",
        "tags": [
          "AI 问答"
        ]
      }
    },
    "/ai/chat": {
      "post": {
        "operationId": "AiController_chat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 事件流 (meta -> delta* -> usage -> done)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "AI 问答"
        ]
      }
    },
...
    "/uploads": {
      "post": {
        "operationId": "UploadController_uploadFiles",
        "parameters": [],
        "responses": {
          "201": {
            "description": "附件上传成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "上传附件（多文件）",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/uploads/{id}": {
      "get": {
        "operationId": "UploadController_getMetadata",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "附件元数据"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取附件元数据",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/uploads/{id}/download": {
      "get": {
        "operationId": "UploadController_downloadFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "附件文件流"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "下载附件文件流",
        "tags": [
          "附件管理"
        ]
      }
    }
  },
  "info": {
    "title": "Chat Server API",
    "description": "聊天服务端 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ModelCatalogItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "模型全局唯一标识",
            "example": "deepseek-chat"
          },
          "label": {
            "type": "string",
            "description": "模型展示名称",
            "example": "DeepSeek Chat"
          },
          "provider": {
            "type": "string",
            "description": "模型归属供应商标识",
            "example": "deepseek"
          }
        },
        "required": [
          "id",
          "label",
          "provider"
        ]
      },
      "ChatRequestDto": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string",
            "description": "会话全局唯一标识。若为空则隐式创建新会话",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "modelId": {
            "type": "string",
            "description": "指定的模型唯一标识",
            "example": "deepseek-chat"
          },
          "content": {
            "type": "string",
            "description": "用户提问正文内容",
            "example": "请解释一下量子纠缠"
          },
          "attachmentIds": {
            "description": "随本次提问携带的附件 ID 列表",
            "example": [
              "123e4567-e89b-12d3-a456-426614174000"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "modelId",
          "content"
        ]
      },
...
    }
  }
}